Backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.8.0
email-validator>=2.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
python-docx>=1.1.0
pdfplumber>=0.10.3

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pymongo>=4.6.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# CORS (built into FastAPI)

# Student Engine Dependencies
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
transformers>=4.30.0
numpy>=1.24.0

# LLM & Vector Search
groq>=0.9.0
qdrant-client>=1.11.0
tenacity>=8.2.0

